+++ packages/veui/demo/cases/Select.vue
@@ -14,10 +14,10 @@
       <veui-select v-bind="attrs" v-model="defaultValue1" disabled></veui-select>
     </section>
     <section>
-      <h2>Slot 样式：</h2>
+      <h2>Slot[name=option-label] 样式：</h2>
       <veui-select v-bind="attrs" v-model="defaultValue3" ui="alt">
-        <template slot="option" scope="props">
-          {{ props.label }}
+        <template slot="option-label" scope="props">
+          <span class="veui-option-custom">{{ props.label }}</span>
         </template>
       </veui-select>
     </section>
@@ -50,7 +50,7 @@
       <h2>Slot 分组样式 1：</h2>
       <veui-select v-bind="optGroupAttrs" v-model="defaultValue7">
         <template slot="option" scope="props">
-          {{ props.label }}
+          <div class="veui-option-custom">{{ props.label }}</div>
         </template>
       </veui-select>
     </section>
@@ -60,7 +60,7 @@
           position: 'bottom right'
         }">
         <template slot="option-label" scope="props">
-          <span class="veui-option-label-text">{{ props.label }}</span>
+          <span class="veui-option-label-text veui-option-custom-label">{{ props.label }}</span>
           <icon name="gift"></icon>
         </template>
       </veui-select>